+++ tools/tui-core/club3090_tui_core/widgets/__init__.py
@@ -1,0 +1,1 @@
+"""Shared Textual widgets for club-3090 TUI apps."""
